third_party/WebKit/LayoutTests/svg/transforms/transform-boxsize-usecounter-no-trigger-1.html

<!DOCTYPE html>
<title>SVG transform: use counter does not trigger</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
rect {
  transform-origin: 100 100;
  translate: -50 -50;
  transform: translate(-50, -50);
}
</style>
<svg>
  <rect width="100" height="100" fill="green" transform="translate(50 50)"/>
</svg>
<script>
// From UseCounter.h
const TransformUsesBoxSizeOnSVG = 1842;

test(function() {
  document.querySelector('rect').getBoundingClientRect();
  assert_false(internals.isUseCounted(document, TransformUsesBoxSizeOnSVG));
});
</script>
